PYTHON-3414 Improve error message when using incompatible pymongocrypt version
