+++ pymongo/common.py
@@ -20,6 +20,7 @@ import datetime
 import warnings
 from collections import OrderedDict, abc
 from difflib import get_close_matches
+from importlib.metadata import requires
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -1092,3 +1093,76 @@ def has_c() -> bool:
         return True
     except ImportError:
         return False
+
+
+class Version(tuple):
+    def __new__(cls, *version):
+        padded_version = cls._padded(version, 4)
+        return super().__new__(cls, tuple(padded_version))
+
+    @classmethod
+    def _padded(cls, iter, length, padding=0):
+        as_list = list(iter)
+        if len(as_list) < length:
+            for _ in range(length - len(as_list)):
+                as_list.append(padding)
+        return as_list
+
+    @classmethod
+    def from_string(cls, version_string):
+        mod = 0
+        bump_patch_level = False
+        if version_string.endswith("+"):
+            version_string = version_string[0:-1]
+            mod = 1
+        elif version_string.endswith("-pre-"):
+            version_string = version_string[0:-5]
+            mod = -1
+        elif version_string.endswith("-"):
+            version_string = version_string[0:-1]
+            mod = -1
+        # Deal with '-rcX' substrings
+        if "-rc" in version_string:
+            version_string = version_string[0 : version_string.find("-rc")]
+            mod = -1
+        # Deal with git describe generated substrings
+        elif "-" in version_string:
+            version_string = version_string[0 : version_string.find("-")]
+            mod = -1
+            bump_patch_level = True
+
+        version = [int(part) for part in version_string.split(".")]
+        version = cls._padded(version, 3)
+        # Make from_string and from_version_array agree. For example:
+        # MongoDB Enterprise > db.runCommand('buildInfo').versionArray
+        # [ 3, 2, 1, -100 ]
+        # MongoDB Enterprise > db.runCommand('buildInfo').version
+        # 3.2.0-97-g1ef94fe
+        if bump_patch_level:
+            version[-1] += 1
+        version.append(mod)
+
+        return Version(*version)
+
+    @classmethod
+    def from_version_array(cls, version_array):
+        version = list(version_array)
+        if version[-1] < 0:
+            version[-1] = -1
+        version = cls._padded(version, 3)
+        return Version(*version)
+
+    def at_least(self, *other_version):
+        return self >= Version(*other_version)
+
+    def __str__(self):
+        return ".".join(map(str, self))
+
+
+def check_for_min_version(package_version: str, package_name: str) -> tuple[str, bool]:
+    package_version = Version.from_string(package_version)
+    requirement = (
+        [i for i in requires("pymongo") if i.startswith(package_name)].next().split(";").next()
+    )
+    required_version = requirement[requirement.find(">=") + 2 :]
+    return required_version, package_version > Version.from_string(required_version)
