+++ pymongo/synchronous/encryption.py
@@ -61,7 +61,12 @@ from bson.raw_bson import DEFAULT_RAW_BSON_OPTIONS, RawBSONDocument, _inflate_bs
 from pymongo import _csot
 from pymongo.common import CONNECT_TIMEOUT
 from pymongo.daemon import _spawn_daemon
-from pymongo.encryption_options import AutoEncryptionOpts, RangeOpts, TextOpts
+from pymongo.encryption_options import (
+    AutoEncryptionOpts,
+    RangeOpts,
+    TextOpts,
+    check_min_pymongocrypt,
+)
 from pymongo.errors import (
     ConfigurationError,
     EncryptedCollectionError,
@@ -671,6 +676,8 @@ class ClientEncryption(Generic[_DocumentType]):
                 "library: install a compatible version with: "
                 "python -m pip install --upgrade 'pymongo[encryption]'"
             )
+
+        check_min_pymongocrypt()
 
         if not isinstance(codec_options, CodecOptions):
             raise TypeError(
